server/requirements.txt
# Main application dependencies
apache-beam[gcp]==2.57.0
pydantic==2.6.4
requests==2.31.0

# Core dependencies with specific versions to avoid conflicts
protobuf>=3.20.2,<4.25.0
grpcio>=1.48.0,<1.59.0
grpcio-status>=1.48.0,<1.59.0

# GCP libraries with compatible versions
google-cloud-bigquery==3.11.4
google-cloud-firestore==2.11.1
google-cloud-language==2.9.1
google-cloud-storage==2.16.0
google-cloud-pubsub==2.18.4

# Core Google libraries - pin to compatible versions
google-api-core==2.15.0
google-auth==2.26.1
googleapis-common-protos==1.56.4
google-cloud-core==2.3.3
google-resumable-media==2.6.0
google-crc32c==1.5.0

# Beam-specific dependencies
dill==0.3.1.1
fastavro==1.9.4
hdfs==2.7.3
httplib2==0.22.0
numpy==1.24.4
orjson>=3.9.15
pyarrow==14.0.2
pymongo>=4.6.3
python-dateutil==2.8.2
pytz==2024.1
regex==2023.12.25
typing-extensions==4.9.0

# Additional utilities
PyYAML==6.0.1
click==8.1.7
python-dotenv==1.0.0
